+++ Constraint/NodeLine.h
@@ -14,22 +14,42 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  ******************************************************************************/
+/**
+ * @class NodeLine
+ * @brief A NodeLine class.
+ *
+ * The `NodeLine` constraint.
+ *
+ * @author tlc
+ * @date 02/05/2021
+ * @version 0.1.0
+ * @file NodeLine.h
+ * @addtogroup Constraint
+ * @{
+ */
 
-#ifndef CONSTRAINTPARSER_H
-#define CONSTRAINTPARSER_H
+#ifndef NODELINE_H
+#define NODELINE_H
 
-#include <suanPan.h>
+#include <Constraint/Constraint.h>
 
-int create_new_constraint(const shared_ptr<DomainBase>&, istringstream&);
+class NodeLine final : public Constraint {
+	static const mat rotation;
 
-void new_embed2d(unique_ptr<Constraint>&, istringstream&);
-void new_fixedlength(unique_ptr<Constraint>&, istringstream&, unsigned);
-void new_minimumgap(unique_ptr<Constraint>&, istringstream&, unsigned);
-void new_mpc(unique_ptr<Constraint>&, istringstream&);
-void new_particlecollision2d(unique_ptr<Constraint>&, istringstream&);
-void new_particlecollision3d(unique_ptr<Constraint>&, istringstream&);
-void new_rigidwall(unique_ptr<Constraint>&, istringstream&, bool, bool);
-void new_bc(unique_ptr<Constraint>&, istringstream&, bool, bool);
-void new_nodeline(unique_ptr<Constraint>&, istringstream&);
+	std::vector<vec> get_position(const shared_ptr<DomainBase>&);
+public:
+	NodeLine(unsigned, unsigned, unsigned, uvec&&);
+
+	int initialize(const shared_ptr<DomainBase>&) override;
+
+	int process(const shared_ptr<DomainBase>&) override;
+
+	void update_status(const vec&) override;
+	void commit_status() override;
+	void clear_status() override;
+	void reset_status() override;
+};
 
 #endif
+
+//! @}